refactor ContentRegistry::ForwardEvent API to ImHexApi::Messaging
